+++ DirectShowFilters/MPUrlSourceSplitter/MPUrlSourceSplitter/MPUrlSourceSplitter_Parser_MSHS/MPUrlSourceSplitter_Parser_MSHS.h
@@ -79,7 +79,7 @@ public:
 
 protected:
 
-  unsigned int lastReceivedLength;
+  size_t lastReceivedLength;
 
   /* methods */
