+++ DirectShowFilters/MPUrlSourceSplitter/MPUrlSourceSplitter/MPUrlSourceSplitter_Protocol_M3U8/MPUrlSourceSplitter_Protocol_M3u8.h
@@ -189,8 +189,8 @@ protected:
   unsigned int lastMediaPlaylistUpdateTime;
 
   // holds last processed size from last store time
-  unsigned int lastProcessedSize;
-  unsigned int currentProcessedSize;
+  size_t lastProcessedSize;
+  size_t currentProcessedSize;
 
   /* methods */
