+++ DirectShowFilters/MPUrlSourceSplitter/MPUrlSourceSplitter/MPUrlSourceSplitter_Protocol_Udp/MPUrlSourceSplitter_Protocol_Udp.h
@@ -166,8 +166,8 @@ protected:
   unsigned int lastReceiveDataTime;
 
   // holds last processed size from last store time
-  unsigned int lastProcessedSize;
-  unsigned int currentProcessedSize;
+  size_t lastProcessedSize;
+  size_t currentProcessedSize;
 
   /* methods */
